Extract findJsModulesFor() to its own module
The Importer class has grown too large. I think we will end up with more
readable, maintainable, and testable code by breaking it into smaller
pieces.

I kept this commit to a minimal change to make reviewing easier. There
may be further opportunities to break down this module into smaller
functions.
