+++ lib/Importer.js
@@ -1,20 +1,15 @@
 'use strict';
 
-const childProcess = require('child_process');
-
 const StringScanner = require('StringScanner');
-const escapeRegExp = require('lodash.escaperegexp');
 const eslint = require('eslint');
 const lodashRange = require('lodash.range');
-const minimatch = require('minimatch');
-const sortBy = require('lodash.sortby');
-const uniqBy = require('lodash.uniqby');
 
 const CommandLineEditor = require('./CommandLineEditor');
 const Configuration = require('./Configuration');
 const ImportStatement = require('./ImportStatement');
 const ImportStatements = require('./ImportStatements');
 const JsModule = require('./JsModule');
+const findJsModulesFor = require('./findJsModulesFor');
 const xRegExp = require('./xregexp');
 
 const REGEX_SKIP_SECTION = xRegExp(`
@@ -83,7 +78,11 @@ class Importer {
   }
 
   goto(variableName) {
-    const jsModules = this.findJsModulesFor(variableName);
+    const jsModules = findJsModulesFor(
+      this.config,
+      this.pathToCurrentFile,
+      variableName
+    );
 
     const jsModule = this.resolveModuleUsingCurrentImports(
       jsModules, variableName);
@@ -212,7 +211,7 @@ class Importer {
     oldImports.imports.forEach((imp) => {
       imp.variables().forEach((variable) => {
         const jsModule = this.resolveModuleUsingCurrentImports(
-          this.findJsModulesFor(variable),
+          findJsModulesFor(this.config, this.pathToCurrentFile, variable),
           variable
         );
 
@@ -255,7 +254,12 @@ class Importer {
    * @return {?JsModule}
    */
   findOneJsModule(variableName) {
-    const jsModules = this.findJsModulesFor(variableName);
+    const jsModules = findJsModulesFor(
+      this.config,
+      this.pathToCurrentFile,
+      variableName
+    );
+
     if (!jsModules.length) {
       return null;
     }
@@ -359,119 +363,6 @@ class Importer {
   }
 
   /**
-   * @param {String} variableName
-   * @return {Array}
-   */
-  findJsModulesFor(variableName) {
-    const aliasModule = this.config.resolveAlias(
-      variableName, this.pathToCurrentFile);
-    if (aliasModule) {
-      return [aliasModule];
-    }
-
-    const namedImportsModule = this.config.resolveNamedExports(variableName);
-    if (namedImportsModule) {
-      return [namedImportsModule];
-    }
-
-    const formattedVarName = this.formattedToRegex(variableName);
-    const egrepCommand =
-      `egrep -i \"(/|^)${formattedVarName}(/index)?(/package)?\.js.*\"`;
-    let matchedModules = [];
-    this.config.get('lookup_paths').forEach((lookupPath) => {
-      if (lookupPath === '') {
-        // If lookupPath is an empty string, the `find` command will not work
-        // as desired so we bail early.
-        throw new Error(`lookup path cannot be empty (${lookupPath})`);
-      }
-
-      const findCommand = [
-        `find ${lookupPath}`,
-        '-name "**.js*"',
-        '-not -path "./node_modules/*"',
-      ].join(' ');
-      const command = `${findCommand} | ${egrepCommand}`;
-
-      // TODO switch to spawn so we can start processing the stream as it comes
-      // in.
-      let out = '';
-      let err = '';
-      try {
-        out = String(childProcess.execSync(command));
-      } catch (error) {
-        err = String(error.stderr);
-      }
-
-      if (err !== '') {
-        throw new Error(err);
-      }
-
-      out.split('\n').forEach((f) => {
-        // TODO: it looks like we process empty strings here too (f === '')
-        if (this.config.get('excludes').some(
-          (globPattern) => minimatch(f, globPattern))) {
-          return;
-        }
-
-        const module = JsModule.construct({
-          lookupPath,
-          relativeFilePath: f,
-          stripFileExtensions:
-            this.config.get('strip_file_extensions', { fromFile: f }),
-          makeRelativeTo:
-            this.config.get('use_relative_paths', { fromFile: f }) &&
-            this.pathToCurrentFile,
-          stripFromPath:
-            this.config.get('strip_from_path', { fromFile: f }),
-        });
-        if (module) {
-          matchedModules.push(module);
-        }
-      });
-    });
-
-    // Find imports from package.json
-    const ignorePrefixes = this.config.get('ignore_package_prefixes').map(
-      (prefix) => escapeRegExp(prefix));
-
-    const depRegex = RegExp(
-      `^(?:${ignorePrefixes.join('|')})?${formattedVarName}$`);
-
-    this.config.packageDependencies().forEach((dep) => {
-      if (!dep.match(depRegex)) {
-        return;
-      }
-
-      const jsModule = JsModule.construct({
-        lookupPath: 'node_modules',
-        relativeFilePath: `node_modules/${dep}/package.json`,
-        stripFileExtensions: [],
-      });
-
-      if (jsModule) {
-        matchedModules.push(jsModule);
-      }
-    });
-
-    this.config.environmentCoreModules().forEach((dep) => {
-      if (dep.toLowerCase() !== variableName.toLowerCase()) {
-        return;
-      }
-
-      matchedModules.push(new JsModule({ importPath: dep }));
-    });
-
-    // If you have overlapping lookup paths, you might end up seeing the same
-    // module to import twice. In order to dedupe these, we remove the module
-    // with the longest path
-    matchedModules = sortBy(matchedModules,
-      (module) => module.importPath.length);
-    matchedModules = uniqBy(matchedModules,
-      (module) => module.filePath);
-    return sortBy(matchedModules, (module) => module.displayName());
-  }
-
-  /**
    * @param {Array} jsModules
    * @param {String} variableName
    * @return {JsModule}
@@ -540,43 +431,6 @@ class Importer {
 
     // Fall back to asking the user to resolve the ambiguity
     return this.resolveOneJsModule(jsModules, variableName);
-  }
-
-  /**
-   * Takes a string in any of the following four formats:
-   *   dash-separated
-   *   snake_case
-   *   camelCase
-   *   PascalCase
-   * and turns it into a star-separated lower case format, like so:
-   *   star*separated
-   *
-   * @param {String} string
-   * @return {String}
-   */
-  formattedToRegex(string) {
-    // Based on
-    // http://stackoverflow.com/questions/1509915/converting-camel-case-to-underscore-case-in-ruby
-
-    // The pattern to match in between words. The "es" and "s" match is there
-    // to catch pluralized folder names. There is a risk that this is overly
-    // aggressive and will lead to trouble down the line. In that case, we can
-    // consider adding a configuration option to control mapping a singular
-    // variable name to a plural folder name (suggested by @lencioni in #127).
-    // E.g.
-    //
-    // {
-    //   "^mock": "./mocks/"
-    // }
-    const splitPattern = '(es|s)?.?';
-
-    // Split up the string, allow pluralizing and a single (any) character
-    // in between. This will make e.g. 'fooBar' match 'foos/bar', 'foo_bar',
-    // and 'foobar'.
-    return string
-      .replace(/([a-z\d])([A-Z])/g, `$1${splitPattern}$2`) // camelCase
-      .replace(/[-_]/g, splitPattern)
-      .toLowerCase();
   }
 }
